# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/warehouse/pick/start": {
      "post": {
        "operationId": "WarehouseController_startPick",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehousePackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "warehouse"
        ]
      }
    },
    "/warehouse/scan": {
      "post": {
        "operationId": "WarehouseController_scan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehousePackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "warehouse"
        ]
      }
    },
    "/warehouse/pack": {
      "post": {
        "operationId": "WarehouseController_pack",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehousePackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "warehouse"
        ]
      }
    },
    "/warehouse/label": {
      "post": {
        "operationId": "WarehouseController_label",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehousePackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "warehouse"
        ]
      }
    },
    "/warehouse/stage": {
      "post": {
        "operationId": "WarehouseController_stage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehousePackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "warehouse"
        ]
      }
    },
...
    "/disponent/assign-driver": {
      "post": {
        "operationId": "DisponentController_assignDriver",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "disponent"
        ]
      }
    },
    "/disponent/reassign-driver": {
      "post": {
        "operationId": "DisponentController_reassignDriver",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "disponent"
        ]
      }
    },
...
    "/driver/pickup/{shipmentId}/complete": {
      "post": {
        "operationId": "DriverController_pickup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickupProofDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ]
      }
    },
...
    "/drivers/pickup/{shipmentId}/complete": {
      "post": {
        "operationId": "DriversController_pickup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickupProofDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ]
      }
    },
...
    "/returns": {
      "post": {
        "operationId": "ReturnsController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      }
    },
    "/returns/{id}/status": {
      "post": {
        "operationId": "ReturnsController_updateStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReturnStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "returns"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI-Agentic Smart Logistics Operating System API",
    "description": "Industrial modular monolith API contract.",
    "version": "0.1.0-rc.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WarehousePackageDto": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          }
        },
        "required": [
          "packageId"
        ]
      },
      "AssignDriverDto": {
        "type": "object",
        "properties": {
          "shipmentId": {
            "type": "string"
          },
          "driverId": {
            "type": "string"
          },
          "packageStatus": {
            "type": "string",
            "enum": [
              "created",
              "reserved",
              "picking",
              "scanned",
              "packed",
              "label_generated",
              "staged",
              "ready_for_dispatch",
              "dispatched",
              "delivered",
              "returned",
              "quarantined",
              "damaged",
              "lost"
            ],
            "default": "ready_for_dispatch"
          }
        },
        "required": [
          "shipmentId",
          "driverId"
        ]
      },
      "PickupProofDto": {
        "type": "object",
        "properties": {
          "packageScanCode": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          },
          "photoObjectKey": {
            "type": "string"
          }
        }
      },
...
      "CreateReturnDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "shipmentId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "customerId",
          "reason"
        ]
      },
      "UpdateReturnStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "return_requested",
              "return_approved",
              "return_rejected",
              "return_pickup_planned",
              "return_picked_up",
              "return_received",
              "inspection_pending",
              "refund_pending",
              "refund_completed",
              "restocked",
              "discarded",
              "closed"
            ]
          },
          "inspection": {
            "type": "object"
          },
          "refundId": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}